+++ src/rattlesnake/environment/modal_environment.py
@@ -38,11 +38,8 @@ import scipy.signal as sig
 from qtpy import QtWidgets, uic
 from qtpy.QtCore import Qt
 
-from rattlesnake.environment.abstract_environment import (
-    AbstractEnvironment,
-    AbstractMetadata,
-    AbstractUI,
-)
+from rattlesnake.environment.abstract_environment import AbstractEnvironment, AbstractMetadata
+from rattlesnake.user_interface.abstract_user_interface import AbstractUI
 from rattlesnake.components.environments import (
     ControlTypes,
     environment_definition_ui_paths,
