+++ src/rattlesnake/environment/random_vibration_sys_id_environment.py
@@ -45,8 +45,8 @@ from qtpy.QtGui import QColor  # pylint: disable=no-name-in-module
 from rattlesnake.environment.abstract_sysid_environment import (
     AbstractSysIdEnvironment,
     AbstractSysIdMetadata,
-    AbstractSysIdUI,
 )
+from rattlesnake.user_interface.abstract_sys_id_user_interface import AbstractSysIdUI
 from rattlesnake.components.environments import (
     ControlTypes,
     environment_definition_ui_paths,
