+++ src/rattlesnake/environment/skeleton_sys_id_environment.py
@@ -33,8 +33,8 @@ from qtpy import QtWidgets, uic
 from rattlesnake.environment.abstract_sysid_environment import (
     AbstractSysIdEnvironment,
     AbstractSysIdMetadata,
-    AbstractSysIdUI,
 )
+from rattlesnake.user_interface.abstract_sys_id_user_interface import AbstractSysIdUI
 from rattlesnake.components.environments import (
     ControlTypes,
     environment_definition_ui_paths,
